refactor(value): Allow any value to have find applied
