+++ crates/value/src/object.rs
@@ -3,11 +3,9 @@ use std::collections::HashMap;
 use std::fmt;
 
 use kstring::KStringCow;
-use liquid_error::{Error, Result};
 
 use crate::map;
 use crate::DisplayCow;
-use crate::PathRef;
 use crate::State;
 use crate::{Value, ValueView};
 
@@ -30,86 +28,6 @@ pub trait ObjectView: ValueView {
     fn contains_key(&self, index: &str) -> bool;
     /// Access a contained `Value`.
     fn get<'s>(&'s self, index: &str) -> Option<&'s dyn ValueView>;
-
-    /// Find a `ValueView` nested in an `ObjectView`
-    #[inline]
-    fn try_find<'o>(&'o self, path: PathRef<'_, '_>) -> Option<&'o dyn ValueView> {
-        let mut indexes = path.iter();
-        let key = indexes.next()?;
-        let key = key.to_kstr();
-        let value = self.get(key.as_str())?;
-
-        indexes.fold(Some(value), |value, index| {
-            let value = value?;
-            if let Some(arr) = value.as_array() {
-                if let Some(index) = index.to_integer() {
-                    arr.get(index)
-                } else {
-                    match &*index.to_kstr() {
-                        "first" => arr.first(),
-                        "last" => arr.last(),
-                        _ => None,
-                    }
-                }
-            } else if let Some(obj) = value.as_object() {
-                obj.get(index.to_kstr().as_str())
-            } else {
-                None
-            }
-        })
-    }
-
-    /// Find a `ValueView` nested in an `ObjectView`
-    #[inline]
-    fn find<'o>(&'o self, path: PathRef<'_, '_>) -> Result<&'o dyn ValueView> {
-        if let Some(res) = self.try_find(path) {
-            Ok(res)
-        } else {
-            for cur_idx in 1..path.len() {
-                let subpath_end = path.len() - cur_idx;
-                let subpath = &path[0..subpath_end];
-                if let Some(parent) = self.try_find(subpath) {
-                    let subpath = itertools::join(subpath.iter().map(ValueView::render), ".");
-                    let requested = &path[subpath_end];
-                    let available = if let Some(arr) = parent.as_array() {
-                        let mut available = vec![
-                            KStringCow::from_static("first"),
-                            KStringCow::from_static("last"),
-                        ];
-                        if 0 < arr.size() {
-                            available.insert(
-                                0,
-                                KStringCow::from_string(format!("0..{}", arr.size() - 1)),
-                            );
-                        }
-                        available
-                    } else if let Some(obj) = parent.as_object() {
-                        let available: Vec<_> = obj.keys().collect();
-                        available
-                    } else {
-                        Vec::new()
-                    };
-                    let available = itertools::join(available.iter(), ", ");
-                    return Error::with_msg("Unknown index")
-                        .context("variable", subpath)
-                        .context("requested index", format!("{}", requested.render()))
-                        .context("available indexes", available)
-                        .into_err();
-                }
-            }
-
-            let requested = path
-                .get(0)
-                .expect("`Path` guarantees at least one element")
-                .to_kstr()
-                .into_owned();
-            let available = itertools::join(self.keys(), ", ");
-            Error::with_msg("Unknown variable")
-                .context("requested variable", requested)
-                .context("available variables", available)
-                .into_err()
-        }
-    }
 }
 
 /// Type representing a Liquid object, payload of the `Value::Object` variant
