+++ crates/value/src/path.rs
@@ -2,6 +2,8 @@ use std::fmt;
 use std::slice;
 
 use itertools;
+use kstring::KStringCow;
+use liquid_error::{Error, Result};
 
 use crate::ScalarCow;
 use crate::ValueView;
@@ -113,3 +115,70 @@ impl<'i, 's: 'i> ExactSizeIterator for PathIter<'i, 's> {
 
 /// Path to a value in an `Object`.
 pub type PathRef<'p, 's> = &'p [ScalarCow<'s>];
+
+/// Find a `ValueView` nested in an `ObjectView`
+pub fn try_find<'o>(value: &'o dyn ValueView, path: PathRef<'_, '_>) -> Option<&'o dyn ValueView> {
+    let indexes = path.iter();
+    indexes.fold(Some(value), |value, index| {
+        let value = value?;
+        if let Some(arr) = value.as_array() {
+            if let Some(index) = index.to_integer() {
+                arr.get(index)
+            } else {
+                match &*index.to_kstr() {
+                    "first" => arr.first(),
+                    "last" => arr.last(),
+                    _ => None,
+                }
+            }
+        } else if let Some(obj) = value.as_object() {
+            obj.get(index.to_kstr().as_str())
+        } else {
+            None
+        }
+    })
+}
+
+/// Find a `ValueView` nested in an `ObjectView`
+pub fn find<'o>(value: &'o dyn ValueView, path: PathRef<'_, '_>) -> Result<&'o dyn ValueView> {
+    if let Some(res) = try_find(value, path) {
+        Ok(res)
+    } else {
+        for cur_idx in 1..path.len() {
+            let subpath_end = path.len() - cur_idx;
+            let subpath = &path[0..subpath_end];
+            if let Some(parent) = try_find(value, subpath) {
+                let subpath = itertools::join(subpath.iter().map(ValueView::render), ".");
+                let requested = &path[subpath_end];
+                let available = if let Some(arr) = parent.as_array() {
+                    let mut available = vec![
+                        KStringCow::from_static("first"),
+                        KStringCow::from_static("last"),
+                    ];
+                    if 0 < arr.size() {
+                        available
+                            .insert(0, KStringCow::from_string(format!("0..{}", arr.size() - 1)));
+                    }
+                    available
+                } else if let Some(obj) = parent.as_object() {
+                    let available: Vec<_> = obj.keys().collect();
+                    available
+                } else {
+                    Vec::new()
+                };
+                let available = itertools::join(available.iter(), ", ");
+                return Error::with_msg("Unknown index")
+                    .context("variable", subpath)
+                    .context("requested index", format!("{}", requested.render()))
+                    .context("available indexes", available)
+                    .into_err();
+            }
+        }
+
+        panic!(
+            "Should have already errored for `{}` with path {:?}",
+            value.source(),
+            path
+        );
+    }
+}
